+++ src/yargs-cmds/discover-hostnames.js
@@ -42,7 +42,7 @@ args to add:
 
 let mod = 
 {
-    // Command name - i.e. gtm-cli <command name> <options>
+    // Command name - i.e. sub-domain-scanner <command name> <options>
     command: "discover-hostnames",
 
     // Command description
@@ -61,7 +61,7 @@ let mod =
     {
         try
         {
-            let allHostnames = [];
+            let allHostnames = new Set();
 
             let domainNames: Array = [];
 
@@ -75,16 +75,21 @@ let mod =
                 domainNames = argv.domainNames;
             }
 
-            for(let hostname of domainNames)
+            for(let domainName of domainNames)
             {
-                const hostnames = await getHostnamesFromCTLogs(hostname);
-                allHostnames = allHostnames.concat(hostnames);
+                const hostnames: Array = await getHostnamesFromCTLogs(domainName);
+
+                for(let hostname of hostnames)
+                {
+                    allHostnames.add(hostname);
+                }
             }
-            
+
             const mustMatch: RegExp = new RegExp(argv.mustMatch);
             const mustNotMatch: RegExp = new RegExp(argv.mustNotMatch);
 
-            const filteredHostnames = filterHostnames(allHostnames, mustMatch, mustNotMatch);
+            const dedupedHostnames = Array.from(allHostnames);
+            const filteredHostnames = filterHostnames(dedupedHostnames, mustMatch, mustNotMatch);
             const output = filteredHostnames.join(EOL);
 
             console.log(output);
@@ -98,5 +103,4 @@ let mod =
     }
 };
 
-// exports.default = mod;
 module.exports = mod;