+++ changelog.md
@@ -6,6 +6,7 @@
 * Dedupe output from discover-hostnames
 * Add wildcard subsitution for common sub-domains for discover-hostnames output
 * Minor fixes and tidies
+* Try to fix Travis (this pkg requires node 11 currently)
 
 ## v1.2.0
 * Add JSON output format for test-hostnames